+++ src/components/page/page.test.js
@@ -6,20 +6,28 @@ import { EuiPage } from './page';
 
 describe('EuiPage', () => {
   test('is rendered', () => {
-    const component = render(
-      <EuiPage {...requiredProps} />
-    );
+    const component = render(<EuiPage {...requiredProps} />);
 
-    expect(component)
-      .toMatchSnapshot();
+    expect(component).toMatchSnapshot();
   });
 
-  test('sets a max-width', () => {
-    const component = render(
-      <EuiPage {...requiredProps} restrictWidth={1024} />
-    );
+  describe('restrict width', () => {
+    test('can be set to a default', () => {
+      const component = render(<EuiPage {...requiredProps} restrictWidth={true} />);
 
-    expect(component)
-      .toMatchSnapshot();
+      expect(component).toMatchSnapshot();
+    });
+
+    test('can be set to a custom number', () => {
+      const component = render(<EuiPage {...requiredProps} restrictWidth={1024} />);
+
+      expect(component).toMatchSnapshot();
+    });
+
+    test('can be set to a custom value and measurement', () => {
+      const component = render(<EuiPage {...requiredProps} restrictWidth="24rem" />);
+
+      expect(component).toMatchSnapshot();
+    });
   });
 });